# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neuraal API",
    "version": "0.1.0",
    "description": "Backend API for Neuraal — task/note management with AI features (summaries, auto-topic classification)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "DevUserIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Temporary dev-only auth. Sends a fixed user ID. Will be replaced by BearerAuth (JWT) in production."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Future JWT auth (not enforced yet)."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Resource not found"
              }
            }
          }
        }
      },
...
      "Attachment": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "entryId",
          "storageKey",
          "filename",
          "mimeType",
          "sizeBytes",
          "kind",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string"
          },
          "entryId": {
            "type": "string",
            "format": "uuid"
          },
          "storageKey": {
            "type": "string"
          },
          "filename": {
            "type": "string",
            "maxLength": 255
          },
          "mimeType": {
            "type": "string",
            "maxLength": 100
          },
          "sizeBytes": {
            "type": "integer"
          },
          "kind": {
            "type": "string",
            "enum": [
              "inline",
              "file"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "ready",
              "deleted"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "parameters": {
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Resource UUID"
      }
    }
  },
  "security": [
    {
      "DevUserIdHeader": []
    }
  ],
  "paths": {
...
    "/api/automations/entry-summary/callback": {
      "post": {
        "tags": [
          "Automations"
        ],
        "summary": "Entry summary callback (from n8n)",
        "operationId": "entrySummaryCallback",
        "description": "Callback endpoint for n8n to deliver AI-generated summaries. Authenticated via HMAC signature (NOT x-user-id).",
        "security": [],
        "parameters": [
          {
            "name": "X-Timestamp",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unix timestamp in ms used for HMAC computation"
          },
          {
            "name": "X-Signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "HMAC-SHA256 signature: hmac(secret, timestamp + '.' + rawBody)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "requestId",
                  "userId",
                  "entryId",
                  "summary",
                  "format"
                ],
                "properties": {
                  "requestId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "entryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "summary": {
                    "type": "string"
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "markdown",
                      "plain"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Summary processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "alreadyProcessed": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/attachments/init": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Initialize attachment upload",
        "operationId": "initAttachmentUpload",
        "description": "Creates an attachment record and returns a presigned PUT URL for uploading.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "entryId",
                  "filename",
                  "mimeType",
                  "sizeBytes",
                  "kind"
                ],
                "properties": {
                  "entryId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "filename": {
                    "type": "string",
                    "maxLength": 255
                  },
                  "mimeType": {
                    "type": "string",
                    "maxLength": 100
                  },
                  "sizeBytes": {
                    "type": "integer"
                  },
                  "kind": {
                    "type": "string",
                    "enum": [
                      "inline",
                      "file"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload initialized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "attachment",
                    "presignedPutUrl"
                  ],
                  "properties": {
                    "attachment": {
                      "$ref": "#/components/schemas/Attachment"
                    },
                    "presignedPutUrl": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "413": {
            "description": "Quota exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/attachments/complete": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Complete attachment upload",
        "operationId": "completeAttachmentUpload",
        "description": "Marks an attachment as ready after successful upload to S3.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "attachmentId"
                ],
                "properties": {
                  "attachmentId": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attachment marked as ready",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "attachment"
                  ],
                  "properties": {
                    "attachment": {
                      "$ref": "#/components/schemas/Attachment"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/attachments/{id}/download": {
      "get": {
        "tags": [
          "Attachments"
        ],
        "summary": "Get attachment download URL",
        "operationId": "getAttachmentDownloadUrl",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "responses": {
          "200": {
            "description": "Presigned download URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "presignedGetUrl"
                  ],
                  "properties": {
                    "presignedGetUrl": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  }
}